Just a test commit. Went back to original

+++ AlphaTracks.cpp
@@ -102,7 +102,7 @@ int main() {
     ifstream SRIMEnergyLossFile("energy_loss_He_11MeV.txt");
     double energyLoss, distanceTravel;
 
-    vector<float> dEdx;
+    vector<double> dEdx;
     vector<double> alphaDistanceTravel;
     vector<double> reverseEnergyLoss;
     vector<double> reverseAlphaDistance;
